refactor: change tabs from SPA to React Router-based routing
