+++ src/pages/index.ts
@@ -1,5 +1,9 @@
 export { default as Home } from './home/Home';
-export { default as SellerProfile } from './user/SellerProfile';
+
+export { default as SelectionTab } from './user/sellerProfile/SelectionTab';
+export { default as ReviewTab } from './user/sellerProfile/ReviewTab';
+export { default as SellerProfile } from './user/sellerProfile/SellerProfile';
+
 export { default as SellerMyPage } from './seller/SellerMyPage';
 
 export { default as NotFound } from './error/NotFound';
