+++ src/pages/user/sellerProfile/ReviewTab.tsx
@@ -1,0 +1,11 @@
+const ReviewTab = () => {
+  return (
+    <section className="flex flex-1">
+      <span className="text-grey06 body-2-m flex w-full justify-center pt-[5.8125rem]">
+        준비 중입니다.
+      </span>
+    </section>
+  );
+};
+
+export default ReviewTab;
