+++ src/routes/AppRouter.tsx
@@ -7,7 +7,14 @@ import {
 import { Suspense } from 'react';
 import { PATH } from '@/routes/path';
 import { GlobalLayout, Loading } from '@/components';
-import { NotFound, Home, SellerProfile, SellerMyPage } from '@/pages';
+import {
+  NotFound,
+  Home,
+  SellerProfile,
+  SellerMyPage,
+  SelectionTab,
+  ReviewTab,
+} from '@/pages';
 
 const router = createBrowserRouter([
   {
@@ -36,11 +43,25 @@ const router = createBrowserRouter([
       },
       {
         path: PATH.SELLER_PROFILE.base,
-        element: <Outlet />,
+        element: (
+          <SellerProfile>
+            <Outlet />
+          </SellerProfile>
+        ),
         children: [
           {
             index: true,
-            element: <SellerProfile />,
+            element: (
+              <Navigate to={PATH.SELLER_PROFILE.tabs.selection} replace />
+            ),
+          },
+          {
+            path: PATH.SELLER_PROFILE.tabs.selection,
+            element: <SelectionTab />,
+          },
+          {
+            path: PATH.SELLER_PROFILE.tabs.review,
+            element: <ReviewTab />,
           },
         ],
       },
