+++ src/routes/path.ts
@@ -5,6 +5,10 @@ export const PATH = {
   },
   SELLER_PROFILE: {
     base: '/seller/:sellerId',
+    tabs: {
+      selection: 'selection',
+      review: 'review',
+    },
   },
   SELLER: {
     base: '/my',
